+++ internal/provider/private_endpoint_services_resource.go
@@ -43,7 +43,15 @@ type privateEndpointServicesResource struct {
 const endpointServicesCreateTimeout = time.Hour
 
 var endpointServicesSchema = schema.Schema{
-	MarkdownDescription: "PrivateEndpointServices contains services that allow for private connectivity to the CockroachDB Cloud cluster.",
+	MarkdownDescription: `PrivateEndpointServices contains services that allow for private connectivity to the CockroachDB Cloud cluster.
+	
+### _Warning_: Use of *private endpoints* requires >=v1.7.6
+If you intend to use this provider to provision [private endpoints](https://github.com/cockroachdb/terraform-provider-cockroach/blob/main/docs/resources/private_endpoint_connection.md):
+- [AWS PrivateLink](https://www.cockroachlabs.com/docs/cockroachcloud/aws-privatelink)
+- [GCP Private Service Connect](https://www.cockroachlabs.com/docs/cockroachcloud/connect-to-your-cluster#gcp-private-service-connect)
+- [Azure Private Link](https://www.cockroachlabs.com/docs/cockroachcloud/cockroachdb-dedicated-on-azure)
+	
+You must install/upgrade to [version 1.7.6](https://github.com/cockroachdb/terraform-provider-cockroach/releases/tag/v1.7.6) or [later](https://registry.terraform.io/providers/cockroachdb/cockroach/latest)`,
 	Attributes: map[string]schema.Attribute{
 		"cluster_id": schema.StringAttribute{
 			Required: true,